requirements.txt
streamlit
pandas
pdfminer.six
python-docx
spacy
scikit-learn
matplotlib
seaborn
